wrap docs in POSSettingsCardView
