+++ Modules/Sources/PointOfSale/Presentation/Settings/PointOfSaleSettingsHardwareDetailView.swift
@@ -198,27 +198,9 @@ struct PointOfSaleSettingsHardwareDetailView: View {
                     }
                     .padding()
                 }
-                Button {
-                    showCardReaderDocumentationModal = true
-                } label: {
-                    HStack(spacing: POSSpacing.medium) {
-                        Image(systemName: "doc.text")
-                            .font(.posBodyLargeRegular())
-                            .accessibilityHidden(true)
-                            .renderedIf(!dynamicTypeSize.isAccessibilitySize)
-
-                        VStack(alignment: .leading, spacing: POSPadding.xSmall) {
-                            Text(Localization.cardReaderDocumentationTitle)
-                                .font(.posBodyLargeRegular())
-                                .dynamicTypeSize(...DynamicTypeSize.accessibility2)
-                            Text(Localization.cardReaderDocumentationSubtitle)
-                                .font(.posBodyMediumRegular())
-                                .foregroundStyle(.secondary)
-                                .dynamicTypeSize(...DynamicTypeSize.accessibility2)
-                        }
-                        Spacer()
-                    }
-                }
+                POSSettingsCardView(title: Localization.cardReaderDocumentationTitle,
+                                    subtitle: Localization.cardReaderDocumentationSubtitle,
+                                    action: { showCardReaderDocumentationModal = true })
                 .padding()
                 .accessibilityAddTraits(.isButton)
                 .listRowSeparator(.hidden)
@@ -461,46 +443,5 @@ private extension PointOfSaleSettingsHardwareDetailView {
 #if DEBUG
 #Preview {
     PointOfSaleSettingsHardwareDetailView(settingsController: PointOfSaleSettingsPreviewController())
-}
-#endif
-
-
-struct POSSettingsCardView: View {
-    let title: String
-    let subtitle: String
-    let action: () -> Void
-
-    @Environment(\.dynamicTypeSize) private var dynamicTypeSize
-
-    var body: some View {
-        Button(action: action) {
-            VStack(alignment: .leading, spacing: POSPadding.xSmall) {
-                Text(title)
-                    .font(.posBodyLargeRegular())
-                    .foregroundStyle(Color.posOnSurface)
-                    .dynamicTypeSize(...DynamicTypeSize.accessibility2)
-                Text(subtitle)
-                    .font(.posBodyMediumRegular())
-                    .foregroundStyle(.secondary)
-                    .dynamicTypeSize(...DynamicTypeSize.accessibility2)
-            }
-            .padding()
-            .frame(maxWidth: .infinity, alignment: .leading)
-            .background(Color.posSurfaceContainerLowest)
-            .posItemCardBorderStyles()
-        }
-        .buttonStyle(.plain)
-        .accessibilityAddTraits(.isButton)
-        .accessibilityLabel("\(title), \(subtitle)")
-    }
-}
-
-#if DEBUG
-#Preview {
-    POSSettingsCardView(
-        title: "Documentation",
-        subtitle: "Learn more about accepting mobile payments",
-        action: { }
-    )
 }
 #endif
